+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2026-04-22
+
+### Fixed
+
+- Sort dev containers last in `wado images` output instead of first
+- Display "dev" instead of raw version string for dev containers in
+  `wado images` output
+
 ## [0.4.5] - 2026-04-22
 
 ### Added
@@ -345,6 +353,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - First release 🎉
+
+[0.5.0]: https://github.com/hpehl/wado/compare/v0.4.5...v0.5.0
 
 [0.4.5]: https://github.com/hpehl/wado/compare/v0.4.4...v0.4.5